move create_dummy_osd to helpers
